requirements.txt
mcp[cli]
openai
python-dotenv
pydantic
rich
pathspec
